uduvgu

+++ 0x12-singly_linked_lists/1-list_len.c
@@ -1,7 +1,7 @@
-#include "lists.h"
 #include <stdlib.h>
 #include <string.h>
 #include <stdio.h>
+#include "lists.h"
 
 /**
  * list_len - find the length of a list
@@ -10,7 +10,7 @@
  */
 size_t list_len(const list_t *h)
 {
-	size_t nodes = 0;
+	int nodes = 0;
 
 	while (h != NULL)
 	{
